new notebook for XGB inspection
